requirements.txt
django-bakery
django
python-decouple
Jinja2~=3.1.2
click~=8.1.3
livereload
python-dotenv~=1.0.0
pytest
Markdown~=3.3.7
PyYAML~=6.0